# swagger-api.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/panasonic-cloud/pump-heat-details": {
      "get": {
        "operationId": "getHeatPumpDetails",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetHeatPumpDetailsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Panasonic Cloud"
        ]
      }
    },
    "/api/panasonic-cloud/change-operation-status": {
      "put": {
        "operationId": "changeOperationStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetHeatPumpStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Panasonic Cloud"
        ]
      }
    },
    "/api/panasonic-cloud/scheduled-water-heating": {
      "put": {
        "operationId": "scheduledWaterHeating",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduledWaterHeatingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Panasonic Cloud"
        ]
      }
    },
    "/api/panasonic-cloud/scheduled-water-heating-status": {
      "get": {
        "operationId": "getScheduledWaterHeatingStatus",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetScheduledWaterHeatingStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Panasonic Cloud"
        ]
      }
    },
    "/api/panasonic-cloud/heating-preferences": {
      "get": {
        "operationId": "getHeatingPreferences",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetHeatingPreferencesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Panasonic Cloud"
        ]
      },
      "put": {
        "operationId": "setHeatingPreferences",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetHeatingPreferencesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Panasonic Cloud"
        ]
      }
    },
    "/api/panasonic-cloud/heating-preferences/auto": {
      "put": {
        "operationId": "setAutomaticHeating",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAutomaticHeatingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Panasonic Cloud"
        ]
      }
    },
...
    "/api/home-device/all": {
      "post": {
        "operationId": "getAllDevices",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllDeviceFilters"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HomeDeviceDetailsDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Home Device"
        ]
      }
    },
    "/api/home-device/create": {
      "post": {
        "operationId": "createDevice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeviceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Home Device"
        ]
      }
    },
    "/api/home-device/updateDeviceName": {
      "put": {
        "operationId": "updateDeviceName",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeDeviceNameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Home Device"
        ]
      }
    },
    "/api/home-device/delete/{id}": {
      "delete": {
        "operationId": "deleteDevice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Home Device"
        ]
      }
    },
    "/api/home-device/details/{id}": {
      "get": {
        "operationId": "getDeviceDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDeviceDetailsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Home Device"
        ]
      }
    },
    "/api/home-device/switch/status/{id}": {
      "post": {
        "operationId": "setPluginSwitchStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPluginSwitchStatusRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Home Device"
        ]
      }
    },
    "/api/home-device/avg-temperature": {
      "get": {
        "operationId": "getHomeAvgTemperature",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetHomeAvgTemperature"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Home Device"
        ]
      }
    },
    "/api/home-device/are-all-doors-and-windows-closed": {
      "get": {
        "operationId": "areAllDoorsAndWindowsClosed",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Home Device"
        ]
      }
    },
...
    "/api/task/create": {
      "post": {
        "operationId": "createTask",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/task/update/{id}": {
      "put": {
        "operationId": "updateTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/task/delete/{id}": {
      "delete": {
        "operationId": "deleteTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/task/update-status/{id}": {
      "get": {
        "operationId": "setTaskStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "active",
            "required": true,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/task/task-list": {
      "get": {
        "operationId": "getTaskList",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Tasks"
        ]
      }
    }
  },
  "info": {
    "title": "",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ZoneStatusDto": {
        "type": "object",
        "properties": {
          "operationStatus": {
            "type": "number"
          },
          "temparatureNow": {
            "type": "number",
            "nullable": true
          },
          "heatSet": {
            "type": "number"
          }
        },
        "required": [
          "operationStatus",
          "temparatureNow",
          "heatSet"
        ]
      },
      "TankStatusDto": {
        "type": "object",
        "properties": {
          "operationStatus": {
            "type": "number"
          },
          "temparatureNow": {
            "type": "number"
          },
          "heatMax": {
            "type": "number"
          },
          "heatMin": {
            "type": "number"
          },
          "heatSet": {
            "type": "number"
          }
        },
        "required": [
          "operationStatus",
          "temparatureNow",
          "heatMax",
          "heatMin",
          "heatSet"
        ]
      },
      "GetHeatPumpDetailsResponse": {
        "type": "object",
        "properties": {
          "deviceGuid": {
            "type": "string"
          },
          "outdoorNow": {
            "type": "number"
          },
          "waterPressure": {
            "type": "string"
          },
          "zoneStatus": {
            "$ref": "#/components/schemas/ZoneStatusDto"
          },
          "tankStatus": {
            "$ref": "#/components/schemas/TankStatusDto"
          }
        },
        "required": [
          "deviceGuid",
          "outdoorNow",
          "waterPressure",
          "zoneStatus",
          "tankStatus"
        ]
      },
      "SetHeatPumpStatusRequest": {
        "type": "object",
        "properties": {
          "isWaterOn": {
            "type": "boolean"
          },
          "isHeatOn": {
            "type": "boolean"
          },
          "deviceGuid": {
            "type": "string"
          }
        },
        "required": [
          "isWaterOn",
          "isHeatOn",
          "deviceGuid"
        ]
      },
      "ScheduledWaterHeatingRequest": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "active"
        ]
      },
      "GetScheduledWaterHeatingStatusResponse": {
        "type": "object",
        "properties": {
          "isScheduled": {
            "type": "boolean"
          }
        },
        "required": [
          "isScheduled"
        ]
      },
      "GetHeatingPreferencesResponse": {
        "type": "object",
        "properties": {
          "isAutomaticHeat": {
            "type": "boolean"
          },
          "groundFlorTemperatureSensorId": {
            "type": "number",
            "nullable": true
          },
          "firstFlorTemperatureSensorId": {
            "type": "number",
            "nullable": true
          },
          "minTargetTemperature": {
            "type": "number",
            "nullable": true
          },
          "maxTargetTemperature": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "isAutomaticHeat",
          "groundFlorTemperatureSensorId",
          "firstFlorTemperatureSensorId",
          "minTargetTemperature",
          "maxTargetTemperature"
        ]
      },
      "SetHeatingPreferencesRequest": {
        "type": "object",
        "properties": {
          "firstFlorSensorId": {
            "type": "number",
            "nullable": true
          },
          "groundFlorSensorId": {
            "type": "number",
            "nullable": true
          },
          "minTargetTemperature": {
            "type": "number",
            "nullable": true
          },
          "maxTargetTemperature": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "firstFlorSensorId",
          "groundFlorSensorId",
          "minTargetTemperature",
          "maxTargetTemperature"
        ]
      },
      "SetAutomaticHeatingRequest": {
        "type": "object",
        "properties": {
          "isAutomaticHeating": {
            "type": "boolean",
            "nullable": false
          }
        },
        "required": [
          "isAutomaticHeating"
        ]
      },
...
      "GetAllDeviceFilters": {
        "type": "object",
        "properties": {
          "deviceType": {
            "type": "number",
            "nullable": true
          },
          "withTaskAssigned": {
            "type": "boolean",
            "nullable": true
          },
          "isOpen": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
      "HomeDeviceDetailsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "type": {
            "type": "number"
          },
          "homeDeviceId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "isOnline": {
            "type": "boolean"
          },
          "signalStrength": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "type",
          "homeDeviceId",
          "name",
          "isOnline",
          "signalStrength"
        ]
      },
      "CreateDeviceRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "name"
        ]
      },
      "ChangeDeviceNameRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": false
          },
          "deviceName": {
            "type": "string",
            "nullable": false
          }
        },
        "required": [
          "id",
          "deviceName"
        ]
      },
...
      "GetDeviceDetailsResponse": {
        "type": "object",
        "properties": {
          "deviceDetails": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/HomeDeviceDetailsDto"
              }
            ]
          }
        },
        "required": [
          "deviceDetails"
        ]
      },
      "SetPluginSwitchStatusRequest": {
        "type": "object",
        "properties": {
          "isOn": {
            "type": "boolean"
          }
        },
        "required": [
          "isOn"
        ]
      },
      "GetHomeAvgTemperature": {
        "type": "object",
        "properties": {
          "avgTemperature": {
            "type": "number"
          }
        },
        "required": [
          "avgTemperature"
        ]
      },
...
      "CreateTaskRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "from": {
            "format": "date-time",
            "type": "string"
          },
          "atSunset": {
            "type": "boolean"
          },
          "to": {
            "format": "date-time",
            "type": "string"
          },
          "assignedDevices": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "name",
          "assignedDevices"
        ]
      },
      "UpdateTaskRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "from": {
            "format": "date-time",
            "type": "string"
          },
          "to": {
            "format": "date-time",
            "type": "string"
          },
          "assignedDevices": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "name",
          "assignedDevices"
        ]
      },
      "TaskDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "startTime": {
            "format": "date-time",
            "type": "string"
          },
          "atSunset": {
            "type": "boolean"
          },
          "endTime": {
            "format": "date-time",
            "type": "string"
          },
          "assignedDevices": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "startTime",
          "atSunset",
          "endTime",
          "assignedDevices",
          "isActive"
        ]
      }
    }
  }
}